+++ healthcheck/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class HealthcheckAppConfig(AppConfig):
+    name = 'healthcheck'